requirements.txt
altair
numpy
pandas
langchain
langchain-community
langchain_experimental
langchain_openai
python-dotenv
streamlit